requirements.txt
tensorflow
numpy
matplotlib
pandas
opencv-python
scikit-learn
seaborn
streamlit
keras
Pillow